+++ User/app_config.h
@@ -6,7 +6,7 @@
 #define __APP_CONFIG_H
 
 #define APP_NAME              "Dev-Beta"
-#define APP_VERSION           "v0.4.6"
+#define APP_VERSION           "v0.4.7"
 #define APP_UI_NAME           "Astra"
 #define APP_TARGET_NAME       "STM32F103C8T6"
 #define APP_LCD_NAME          "ST7735S_128x160"